leak-detective: Try to properly free allocations after deinitialization
If a function we whitelist allocates memory while leak detective is enabled
but only frees it after LD has already been disabled, free() will get called
with invalid pointers (not accounting for the headers used by LD), which might
cause checks in the C library to fail.  This tries to detect such cases and
calling free with the correct pointer.